refactor(app): route command interface through CLI11 for consistent parsing
